Add files via upload
About the Book"Fundamentals Of IoT" is a comprehensive guide designed to introduce readers to the fascinating world of the Internet of Things (IoT). This book is a practical project manual that provides step-by-step instructions for various IoT projects, making it an invaluable resource for beginners and enthusiasts alike

Authors and Contributors:
This book was created with the support and collaboration of the following individuals:
Dr. Mohammed Ashik (Author)
Akshitha Jangam (Contributor, IIIT Srikakulam)
Lakshmi Nagaraju Kojja (Contributor, IIIT Srikakulam)
Bharath Kumar Lanjapalli (Contributor, IIIT Srikakulam)